+++ src/components/index.ts
@@ -27,7 +27,7 @@ export { default as DragDrop } from './DragDrop/DragDrop'
 export { default as SliderContents } from './SliderContents/SliderContents'
 export { default as MarkdownPreviewer } from './MarkdownPreviewer/MarkdownPreviewer'
 export { default as ComponentList } from './ComponentList/ComponentList'
-export { default as DocumentList } from './DocumentList/DocumentList'
+export { default as DocumentListTree } from './DocumentListTree/DocumentListTree'
 export { default as ComponentPreviewTabs } from './ComponentPreviewTabs/ComponentPreviewTabs'
 export { default as MatrixRain } from './MatrixRain/MatrixRain'
 export { default as MatrixArea } from './MatrixArea/MatrixArea'
