+++ src/pages/DocumentsPreviewer.tsx
@@ -1,8 +1,5 @@
 import React from 'react'
-import TreeView from '@mui/lab/TreeView'
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import ChevronRightIcon from '@mui/icons-material/ChevronRight'
-import { DocumentList, MarkdownPreviewer } from '@/components'
+import { DocumentListTree, MarkdownPreviewer } from '@/components'
 import { Box } from '@mui/material'
 import { useSetParams } from '@/customHooks'
 
@@ -12,14 +9,7 @@ const DocumentsPreviewer: React.FC = () => {
   return (
     <Box sx={{ margin: '2% 2% 5% 2%', display: 'flex' }}>
       <Box sx={{ padding: 3, border: 'solid 1px #06D8D7', borderRadius: 3 }}>
-        <TreeView
-          aria-label="icon expansion"
-          defaultCollapseIcon={<ExpandMoreIcon />}
-          defaultExpandIcon={<ChevronRightIcon />}
-          sx={{ overflowY: 'auto' }}
-        >
-          <DocumentList getParams={getParams} />
-        </TreeView>
+        <DocumentListTree getParams={getParams} />
       </Box>
       <>
         {params ? (
